[wip/pinocchio] Delete: moved to robotpkg
The package was split in math/pinocchio for the C++ library and py-pinocchio
for the python bindings.

Note that at the moment, the headers provided by the python bindings are not
installed because they would conflict between multiple python version. The .pc
file is not installed either for the same reason. If any of those files are
actually needed, one should figure out a proper namespace or a .pc name.